Use DAG-CBOR for hash value instead of canonical JSON
The canonical_json library is thoroughly broken:
https://github.com/mozilla-services/canonicaljson-rs/issues/9

JSON is an extremely difficult format to round-trip, and the canonical
JSON spec has some issues.
(See https://github.com/simon-greatrix/CanonicalJson#asymmetric-attack-via-integer-expansion
for one example).
It is even more difficult to round-trip JSON in statically-typed
languages, and in JavaScript. Most of the several "canonical" JSON specs
punt on the issues that concern us. E.g., by prohibiting floats
entirely.
(https://github.com/engineerd/cjson/blob/7ac51c460748c9bb6dc5f17589d4f2245fca496a/testdata/errors/float.json)

Most other formats have similar piles of inconsistencies
(https://github.com/msgpack/msgpack/issues/215#issuecomment-242991200)

[DAG-CBOR](https://ipld.io/specs/codecs/dag-cbor/spec/) is not perfect
either, and recommends that floats be avoided (for good reason).
But it seems to be the best available format for consistent
serialization, and it has a working rust implementation.

We only use DAG-CBOR for the hash. We still use JSON for passing data
between processes, so we still have to deal with the round-tripping
inconsistencies. We address this by considering the output of the
add-hashes step to be canonical. If a process cannot guarantee identical
output, it should pass the data to the add-hashes step and use its
output. This only means that "canonical" data can be round-tripped. Non-
canonical data may be changed. E.g., a float's value may change if it is
deserialized into a type with more or less precision than its
non-canonical representation. Non-canonical data may also cause an
error, as in the case of non-Unicode strings in JSON.