Run own static file server for tests
webfsd-based tests sometimes failed in CI, possibly because the
server wasn't up before the test ran. It was also limited to unix,
and I was not able to figure out how to shut webfsd down after testing.
This replaces webfsd with our own actix-based file server and waits for
the server to be ready. The web-based tests now run on Windows too.